+++ edaflow/__init__.py
@@ -2,9 +2,9 @@
 edaflow - A Python package for exploratory data analysis workflows
 """
 
-from .analysis import check_null_columns, analyze_categorical_columns, convert_to_numeric, visualize_categorical_values, display_column_types, impute_numerical_median, impute_categorical_mode, visualize_numerical_boxplots
+from .analysis import check_null_columns, analyze_categorical_columns, convert_to_numeric, visualize_categorical_values, display_column_types, impute_numerical_median, impute_categorical_mode, visualize_numerical_boxplots, handle_outliers_median
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"
 __author__ = "Evan Low"
 __email__ = "evan.low@illumetechnology.com"
 
@@ -29,5 +29,10 @@ __all__ = [
     'check_null_columns', 
     'analyze_categorical_columns', 
     'convert_to_numeric', 
-    'visualize_categorical_values'
+    'visualize_categorical_values',
+    'display_column_types',
+    'impute_numerical_median',
+    'impute_categorical_mode',
+    'visualize_numerical_boxplots',
+    'handle_outliers_median'
 ]